@@ -1,29 +1,24 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <TargetFramework>net5.0-windows10.0.19041.0</TargetFramework>
-    <TargetPlatformMinVersion>10.0.17763.0</TargetPlatformMinVersion>
-    <RootNamespace>MaterialDesignThemes.WinUI</RootNamespace>
-    <RuntimeIdentifiers>win10-x86;win10-x64;win10-arm64</RuntimeIdentifiers>
-    <ManagePackageVersionsCentrally>false</ManagePackageVersionsCentrally>
-  </PropertyGroup>
-  <ItemGroup>
-    <None Remove="Themes\MaterialDesignTheme.Defaults.xaml" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <FrameworkReference Update="Microsoft.Windows.SDK.NET.Ref" RuntimeFrameworkVersion="10.0.18362.16" />
-    <FrameworkReference Update="Microsoft.Windows.SDK.NET.Ref" TargetingPackVersion="10.0.18362.16" />
-  </ItemGroup>
+    
+    <PropertyGroup>
+        <TargetFramework>net6.0-windows10.0.19041.0</TargetFramework>
+        <TargetPlatformMinVersion>10.0.17763.0</TargetPlatformMinVersion>
+        <RootNamespace>MaterialDesignThemes.WinUI</RootNamespace>
+        <RuntimeIdentifiers>win10-x86;win10-x64;win10-arm64</RuntimeIdentifiers>
+        <UseWinUI>true</UseWinUI>
+    </PropertyGroup>
 
     <ItemGroup>
-      <PackageReference Include="Microsoft.ProjectReunion" Version="0.5.5" />
-      <PackageReference Include="Microsoft.ProjectReunion.Foundation" Version="0.5.5" />
-      <PackageReference Include="Microsoft.ProjectReunion.WinUI" Version="0.5.5" />
-  </ItemGroup>
-
+        <PackageReference Include="Microsoft.WindowsAppSDK" />
+        <PackageReference Include="Microsoft.Windows.SDK.BuildTools" />
+    </ItemGroup>
+    
     <ItemGroup>
-      <Page Update="Themes\MaterialDesignTheme.Defaults.xaml">
-        <Generator>MSBuild:Compile</Generator>
-      </Page>
+        <None Remove="Themes\MaterialDesignTheme.Defaults.xaml" />
+    </ItemGroup>
+    <ItemGroup>
+        <Page Update="Themes\MaterialDesignTheme.Defaults.xaml">
+            <Generator>MSBuild:Compile</Generator>
+        </Page>
     </ItemGroup>
 </Project>
